refactor: make it a data type
